requirements.txt
apipkg==1.5
appdirs==1.4.4
asgiref==3.2.7
attrs==19.3.0
black==19.10b0
click==7.1.2
coverage==5.1
Django==3.0.6
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
execnet==1.7.1
importlib-metadata==1.6.1
more-itertools==8.4.0
packaging==20.4
pathspec==0.8.0
pluggy==0.13.1
py==1.8.2
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.3
pytest-cov==2.10.0
pytest-django==3.9.0
pytest-forked==1.1.3
pytest-xdist==1.32.0
pytz==2020.1
regex==2020.6.8
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
wcwidth==0.2.4
zipp==3.1.0
